requirements.txt
# Core dependencies (no mitmproxy — works with Python 3.13+)
requests>=2.28.0
python-dotenv>=1.0.0
rich>=13.0.0
colorama>=0.4.6

# Optional — uncomment if you need database storage
# psycopg2-binary>=2.9.7
# pymongo>=4.6.0
# pydantic>=2.5.0
# schedule>=1.2.0